requirements.txt
streamlit
PyPDF2
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
faiss-cpu
tenacity
google-generativeai
deep-translator
langdetect
googletrans==4.0.0-rc1
pdfplumber
pytesseract